KL-176/refactor: seaprate initializeSearchState to util
