+++ src/utils/initializeSearchState.js
@@ -1,0 +1,12 @@
+const initializeSearchState = (searchedCategory, searchedContent) => {
+  const result = {
+    countries: [],
+    cities: [],
+    categories: [],
+    subcategories: [],
+  }
+  result[searchedCategory].push(searchedContent)
+  return result
+}
+
+export default initializeSearchState
